backend/requirements.txt
openai
python-dotenv
pymongo
pydantic
fastapi
uvicorn